+++ src/app/layout.metadata.ts
@@ -1,0 +1,31 @@
+import { Metadata } from 'next'
+
+export const metadata: Metadata = {
+  title: {
+    template: 'ꓘO-∀ | %s',
+    default: 'ꓘO-∀',
+  },
+  description:
+    'Explore our latest projects and developments in AI, web development, and creative coding.',
+  openGraph: {
+    title: 'ꓘO-∀',
+    description:
+      'Explore our latest projects and developments in AI, web development, and creative coding.',
+    images: [
+      {
+        url: 'https://www.apesonkeys.com/a-okay-monkey-1.png',
+        width: 1200,
+        height: 630,
+        alt: 'An Ape On Keys',
+      },
+    ],
+  },
+  twitter: {
+    card: 'summary_large_image',
+    title: 'ꓘO-∀',
+    description:
+      'Explore our latest projects and developments in AI, web development, and creative coding.',
+    images: ['https://www.apesonkeys.com/a-okay-monkey-1.png'],
+    creator: '@apesonkeys',
+  },
+}
