+++ src/app/layout.tsx
@@ -1,9 +1,11 @@
-import type { Metadata } from 'next'
 import localFont from 'next/font/local'
 import './globals.css'
 import { ConvexClientProvider } from './ConvexClientProvider'
 import Navigation from '@/components/Navigation'
 import { VT323 } from 'next/font/google'
+import { metadata } from './layout.metadata'
+
+export { metadata }
 
 const geistSans = localFont({
   src: './fonts/GeistVF.woff',
@@ -20,45 +22,6 @@ const vt323 = VT323({
   subsets: ['latin'],
   variable: '--font-vt323',
 })
-
-export const metadata: Metadata = {
-  title: {
-    template: '%s | An Ape On Keys',
-    default: 'An Ape On Keys',
-  },
-  description:
-    'Explore our latest projects and developments in AI, web development, and creative coding.',
-  metadataBase: new URL('https://www.apesonkeys.com'),
-  openGraph: {
-    type: 'website',
-    url: 'https://www.apesonkeys.com',
-    title: 'An Ape On Keys',
-    description:
-      'Explore our latest projects and developments in AI, web development, and creative coding.',
-    images: [
-      {
-        url: 'https://www.apesonkeys.com/a-okay-monkey-1.png',
-        width: 1200,
-        height: 630,
-        alt: 'An Ape On Keys',
-      },
-    ],
-  },
-  twitter: {
-    card: 'summary_large_image',
-    title: 'An Ape On Keys',
-    description:
-      'Explore our latest projects and developments in AI, web development, and creative coding.',
-    creator: '@apesonkeys',
-    images: ['https://www.apesonkeys.com/a-okay-monkey-1.png'],
-  },
-  icons: {
-    icon: '/favicon.ico',
-    shortcut: '/favicon-16x16.png',
-    apple: '/apple-touch-icon.png',
-  },
-  manifest: '/site.webmanifest',
-}
 
 export default function RootLayout({
   children,
