+++ src/app/metadata.ts
@@ -1,0 +1,6 @@
+import { Metadata } from 'next'
+
+export const metadata: Metadata = {
+  title: 'ꓘO-∀',
+  // ... rest of metadata if needed
+}
